api/requirements.txt
asgiref==3.7.2
Django==5.0
psycopg2==2.9.9
sqlparse==0.4.4
tzdata==2023.3

#Requirements for GraphQL
aniso8601==9.0.1
django-graphql-jwt==0.4.0
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
PyJWT==2.8.0
six==1.16.0
text-unidecode==1.3

# Requirements for testcases
freezegun==1.4.0